backend/src/main/java/com/gabi/customer/CustomerJPADataAccessService.java
package com.gabi.customer;

import org.springframework.jdbc.support.GeneratedKeyHolder;
import org.springframework.stereotype.Repository;

import java.util.List;
import java.util.Optional;
@Repository("jpa")
public class CustomerJPADataAccessService implements CustomerDao{

    private final CustomerRepository customerRepository;
    GeneratedKeyHolder generatedKeyHolder = new GeneratedKeyHolder();

    public CustomerJPADataAccessService(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    @Override
    public List<Customer> selectAllCustomer() {
        return customerRepository.findAll();
    }

    @Override
    public Optional<Customer> getCustomerById(Integer id) {
        return  customerRepository.findById(id);
    }

    @Override
    public void insertCustomer(Customer customer) {
        customerRepository.save(customer);
    }

    @Override
    public boolean existPersonWithEmail(String email) {
        return  customerRepository.existsCustomerByEmail(email);
    }

    @Override
    public boolean existsPersonwithId(Integer id) {
        return customerRepository.existsById(id);
    }

    @Override
    public void deleteCustomerbyID(Integer id) {
        customerRepository.deleteById(id);
    }

    @Override
    public void updateCustomer(Customer customer) {
        customerRepository.save(customer);
    }
}
